Added PacketFetcher

+++ src/de/selebrator/npc/EnumEntityStatus.java
@@ -10,7 +10,7 @@ public enum EnumEntityStatus {
         this.id = id;
     }
 
-    public int getId() {
-        return this.id;
+    public byte getId() {
+        return (byte) this.id;
     }
 }
